fix: moved Friend to new file
