+++ lib/followers/isUserFollowingMultiple.ts
@@ -1,15 +1,11 @@
 import getSpotifyData from "@/lib/getSpotifyData";
 
-const isUserFollowingMultiple = async (usernames: string) => {
-  const response =
-    await getSpotifyData<SpotifyApi.UserFollowsUsersOrArtistsResponse>(
-      `https://api.spotify.com/v1/me/following/contains?type=user&ids=${usernames}`,
-      {
-        cache: "no-cache",
-      }
-    );
-
-  return response;
-};
+const isUserFollowingMultiple = async (usernames: string) =>
+  getSpotifyData<SpotifyApi.UserFollowsUsersOrArtistsResponse>(
+    `https://api.spotify.com/v1/me/following/contains?type=user&ids=${usernames}`,
+    {
+      cache: "no-cache",
+    }
+  );
 
 export default isUserFollowingMultiple;
